refactor playground/** into screens/PlaygroundNavigator
