+++ screens/PlaygroundNavigator/sections/PlaygroundToken.tsx
@@ -2,12 +2,12 @@ import { TokenInfo } from '@defichain/jellyfish-api-core/dist/category/token'
 import { PlaygroundRpcClient } from '@defichain/playground-api-client'
 import React, { useEffect, useState } from 'react'
 import tailwind from 'tailwind-rn'
-import { Text, View } from '../../components'
-import { usePlaygroundRpcClient } from '../../hooks/api/usePlaygroundClient'
-import { useWalletAPI } from '../../hooks/wallet/WalletAPI'
-import { WalletStatus } from '../../store/wallet'
-import { PlaygroundAction } from '../Playground.Action'
-import { PlaygroundStatus } from '../Playground.Status'
+import { Text, View } from '../../../components'
+import { usePlaygroundRpcClient } from '../../../hooks/api/usePlaygroundClient'
+import { useWalletAPI } from '../../../hooks/wallet/WalletAPI'
+import { WalletStatus } from '../../../store/wallet'
+import { PlaygroundAction } from '../components/PlaygroundAction'
+import { PlaygroundStatus } from '../components/PlaygroundStatus'
 
 export function PlaygroundToken (): JSX.Element | null {
   const WalletAPI = useWalletAPI()
