+++ screens/PlaygroundNavigator/sections/PlaygroundUTXO.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from 'react'
 import tailwind from 'tailwind-rn'
-import { Text, View } from '../../components'
-import { usePlaygroundApiClient, usePlaygroundRpcClient } from '../../hooks/api/usePlaygroundClient'
-import { useWalletAPI } from '../../hooks/wallet/WalletAPI'
-import { WalletStatus } from '../../store/wallet'
-import { PlaygroundAction } from '../Playground.Action'
-import { PlaygroundStatus } from '../Playground.Status'
+import { Text, View } from '../../../components'
+import { usePlaygroundApiClient, usePlaygroundRpcClient } from '../../../hooks/api/usePlaygroundClient'
+import { useWalletAPI } from '../../../hooks/wallet/WalletAPI'
+import { WalletStatus } from '../../../store/wallet'
+import { PlaygroundAction } from '../components/PlaygroundAction'
+import { PlaygroundStatus } from '../components/PlaygroundStatus'
 
 export function PlaygroundUTXO (): JSX.Element | null {
   const WalletAPI = useWalletAPI()
